Fix installing ansible 1.9

+++ Vagrantfile
@@ -11,8 +11,11 @@ Vagrant.require_version ">= 1.8.0"
 
 $bootstrap_script = <<SCRIPT
 # install ansible 1.9
-rm /etc/apt/sources.list.d/ansible-ansible-trusty.list 2>/dev/null
-yes|add-apt-repository ppa:ansible/ansible-1.9 2>/dev/null
+if ! hash ansible 2>/dev/null; then
+    rm /etc/apt/sources.list.d/ansible-ansible-trusty.list 2>/dev/null
+    yes|add-apt-repository ppa:ansible/ansible-1.9 2>/dev/null
+    apt-get update && apt-get -y install ansible
+fi
 # install git
 if ! hash git 2>/dev/null; then
     apt-get update && apt-get -y install git
